Feture1647 bring the option to whitelist some rules / rename MethodIsOverridden rule
