+++ tests/data/Evaluate/Core/MethodIsOverridden/FooPlugin.php
@@ -5,14 +5,8 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace TestCore\MethodIsOverwritten;
+namespace TestCore\MethodIsOverridden;
 
-class CoreRepository
+class FooPlugin implements FooPluginInterface
 {
-    /**
-     * @return void
-     */
-    public function queryEntityData(): void
-    {
-    }
 }
