+++ tests/data/Evaluate/Core/MethodIsOverridden/FooPluginInterface.php
@@ -5,7 +5,7 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace TestCore\MethodIsOverwritten;
+namespace TestCore\MethodIsOverridden;
 
 interface FooPluginInterface
 {
